refactor: Move augmentation nodes in individual files
Implements ML-597
